refactor: split into separate GoogleSignInButton and AppleSignInButton files
- Replace SignInButtons.tsx with two dedicated files:
  GoogleSignInButton.tsx and AppleSignInButton.tsx, each in components/buttons/
- Switch Google SVG from the 18x18 scaled-down derivation to the verbatim
  48x48 paths published by Google's own button builder
  (https://developers.google.com/identity/sign-in/web/build-button),
  rendered at 18x18 via viewBox; source explicitly documented in file
- Add Apple HIG link to AppleSignInButton for the same reason
- Update SignInDialog.tsx and Signin.tsx import paths accordingly
