+++ apps/antalmanac/src/components/buttons/GoogleSignInButton.tsx
@@ -1,0 +1,82 @@
+import { useThemeStore } from '$stores/SettingsStore';
+import { Button } from '@mui/material';
+
+/**
+ * Colored four-path Google "G" logo.
+ *
+ * Path data is sourced verbatim from Google's official Sign in with Google
+ * button builder (48 × 48 viewBox):
+ *   https://developers.google.com/identity/sign-in/web/build-button
+ *
+ * Rendered at 18 × 18 display size via the SVG viewBox attribute — the paths
+ * themselves are unchanged from Google's published asset.
+ *
+ * Per Google's branding guidelines, monochrome versions of the "G" are
+ * explicitly disallowed, and the logo must appear on a white (light) or
+ * near-black (dark) background only:
+ *   https://developers.google.com/identity/branding-guidelines
+ */
+const GoogleLogo = () => (
+    <svg width="18" height="18" viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg" aria-hidden="true">
+        {/* Red — top-left arc */}
+        <path
+            fill="#EA4335"
+            d="M24 9.5c3.54 0 6.71 1.22 9.21 3.6l6.85-6.85C35.9 2.38 30.47 0 24 0 14.62 0 6.51 5.38 2.56 13.22l7.98 6.19C12.43 13.72 17.74 9.5 24 9.5z"
+        />
+        {/* Blue — right arc + horizontal bar */}
+        <path
+            fill="#4285F4"
+            d="M46.98 24.55c0-1.57-.15-3.09-.38-4.55H24v9.02h12.94c-.58 2.96-2.26 5.48-4.78 7.18l7.73 6c4.51-4.18 7.09-10.36 7.09-17.65z"
+        />
+        {/* Yellow — bottom-left arc */}
+        <path
+            fill="#FBBC05"
+            d="M10.53 28.59c-.48-1.45-.76-2.99-.76-4.59s.27-3.14.76-4.59l-7.98-6.19C.92 16.46 0 20.12 0 24c0 3.88.92 7.54 2.56 10.78l7.97-6.19z"
+        />
+        {/* Green — bottom arc */}
+        <path
+            fill="#34A853"
+            d="M24 48c6.48 0 11.93-2.13 15.89-5.81l-7.73-6c-2.15 1.45-4.92 2.3-8.16 2.3-6.26 0-11.57-4.22-13.47-9.91l-7.98 6.19C6.51 42.62 14.62 48 24 48z"
+        />
+    </svg>
+);
+
+interface GoogleSignInButtonProps {
+    onClick: () => void;
+    fullWidth?: boolean;
+}
+
+/**
+ * Sign in with Google button styled to Google's branding spec.
+ *
+ * Light theme: white fill (#FFFFFF), #747775 1 px border, #1F1F1F text.
+ * Dark theme:  #131314 fill, #8E918F 1 px border, #E3E3E3 text.
+ * Font: Roboto Medium (per spec), no box-shadow at rest.
+ */
+export const GoogleSignInButton = ({ onClick, fullWidth }: GoogleSignInButtonProps) => {
+    const isDark = useThemeStore((store) => store.isDark);
+    return (
+        <Button
+            onClick={onClick}
+            startIcon={<GoogleLogo />}
+            variant="outlined"
+            size="large"
+            fullWidth={fullWidth}
+            sx={{
+                backgroundColor: isDark ? '#131314' : '#ffffff',
+                color: isDark ? '#e3e3e3' : '#1f1f1f',
+                border: `1px solid ${isDark ? '#8e918f' : '#747775'}`,
+                fontFamily: 'Roboto, sans-serif',
+                fontWeight: 500,
+                boxShadow: 'none',
+                '&:hover': {
+                    backgroundColor: isDark ? '#1e1e1f' : '#f2f2f2',
+                    boxShadow: '0 1px 3px rgba(0,0,0,0.15)',
+                    border: `1px solid ${isDark ? '#8e918f' : '#747775'}`,
+                },
+            }}
+        >
+            Sign in with Google
+        </Button>
+    );
+};
